requirements.txt
future==0.18.3
iso8601==0.1.12
pyserial==3.4
python-periphery==2.3.0
PyYAML==6.0.1
flake8==3.8.3
pytest==7.0.1
pytest-cov==2.8.1
filelock==3.0.12
pydocstyle==5.0.2
packaging>=20.4
pre-commit
